+++ lib/ar_captcha.dart
@@ -2,4 +2,5 @@ library;
 
 export 'res/enums/data_size.dart';
 export 'res/enums/captcha_type.dart';
+export 'res/model/captcha_params.dart';
 export 'controller/ar_captcha_controller.dart';