+++ lib/res/model/captcha_params.dart
@@ -1,0 +1,14 @@
+// Project imports:
+import '../enums/captcha_type.dart';
+
+class CaptchaParams {
+  final CaptchaType mode;
+  final Function(String error) onError;
+  final Function(String token) onSuccess;
+
+  CaptchaParams({
+    required this.mode,
+    required this.onError,
+    required this.onSuccess,
+  });
+}